+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 
+- [!1664](https://gitlab.com/meltano/meltano/-/merge_requests/1664) Add `env` properties to custom plugin `settings` in `meltano.yml`
+
 ### Fixes
 
 ### Breaks